Unify single and multiple-layer MetaLayer models
